Add 'my-web-service-api' subproject - refactor packages' names - refactor classes' names - organize imports
