+++ build.gradle
@@ -7,6 +7,23 @@ plugins {
   id 'maven-publish'
 }
 
+allprojects {
+
+  repositories {
+    mavenCentral()
+  }
+
+  apply plugin: 'java'
+
+  dependencies {
+    compileOnly 'org.projectlombok:lombok:1.18.24'
+    annotationProcessor 'org.projectlombok:lombok:1.18.24'
+
+    testCompileOnly 'org.projectlombok:lombok:1.18.24'
+    testAnnotationProcessor 'org.projectlombok:lombok:1.18.24'
+  }
+}
+
 apply from: "${rootDir}/gradle/test.gradle"
 
 repositories {
@@ -18,6 +35,7 @@ repositories {
 }
 
 dependencies {
+  implementation project(":my-web-service-api")
   implementation(group: 'ch.qos.logback', name: 'logback-core', version: '1.2.10')
 
   implementation(group: 'org.springframework.boot', name: 'spring-boot-starter-web')
@@ -51,7 +69,7 @@ dependencies {
 }
 
 group = 'pl.tchorzyksen'
-version = '0.0.2-SNAPSHOT'
+version = '0.1.1-SNAPSHOT'
 description = 'my-web-service'
 java.sourceCompatibility = JavaVersion.VERSION_17
