+++ my-web-service-api/src/main/java/pl/tchorzyksen/my/web/service/model/dto/PersonDto.java
@@ -1,0 +1,13 @@
+package pl.tchorzyksen.my.web.service.model.dto;
+
+import lombok.Data;
+
+@Data
+public class PersonDto {
+
+  private long id;
+
+  private String firstName;
+
+  private String lastName;
+}
