+++ my-web-service-api/src/main/java/pl/tchorzyksen/my/web/service/model/dto/UserDto.java
@@ -1,0 +1,28 @@
+package pl.tchorzyksen.my.web.service.model.dto;
+
+import java.io.Serial;
+import java.io.Serializable;
+import lombok.Data;
+
+@Data
+public class UserDto implements Serializable {
+
+  @Serial
+  private static final long serialVersionUID = 1L;
+
+  private long id;
+
+  private String userId;
+
+  private PersonDto person;
+
+  private String email;
+
+  private String password;
+
+  private String encryptedPassword;
+
+  private String emailVerificationToken;
+
+  private Boolean emailVerificationStatus = false;
+}
