+++ my-web-service-api/src/main/java/pl/tchorzyksen/my/web/service/model/request/PersonRequest.java
@@ -1,0 +1,11 @@
+package pl.tchorzyksen.my.web.service.model.request;
+
+import lombok.Data;
+
+@Data
+public class PersonRequest {
+
+  private String firstName;
+
+  private String lastName;
+}
