+++ my-web-service-api/src/main/java/pl/tchorzyksen/my/web/service/model/request/UserLoginRequest.java
@@ -1,0 +1,13 @@
+package pl.tchorzyksen.my.web.service.model.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserLoginRequest {
+
+  private String email;
+
+  private String password;
+}
