+++ my-web-service-api/src/main/java/pl/tchorzyksen/my/web/service/model/request/UserRequest.java
@@ -1,0 +1,13 @@
+package pl.tchorzyksen.my.web.service.model.request;
+
+import lombok.Data;
+
+@Data
+public class UserRequest {
+
+  private PersonRequest person;
+
+  private String email;
+
+  private String password;
+}
