+++ my-web-service-api/src/main/java/pl/tchorzyksen/my/web/service/model/response/PersonResponse.java
@@ -1,0 +1,13 @@
+package pl.tchorzyksen.my.web.service.model.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class PersonResponse {
+
+  private String firstName;
+
+  private String lastName;
+}
