+++ my-web-service-api/src/main/java/pl/tchorzyksen/my/web/service/model/response/UserResponse.java
@@ -1,0 +1,15 @@
+package pl.tchorzyksen.my.web.service.model.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserResponse {
+
+  private String userId;
+
+  private PersonResponse person;
+
+  private String email;
+}
