--- src/functional-test/java/pl/tchorzyksen/FunctionalTestConfiguration.java
@@ -1,7 +1,0 @@
-package pl.tchorzyksen;
-
-import org.springframework.boot.test.context.TestConfiguration;
-
-@TestConfiguration
-public class FunctionalTestConfiguration {
-}
