+++ src/functional-test/java/pl/tchorzyksen/my/web/service/FunctionalTestConfiguration.java
@@ -1,0 +1,7 @@
+package pl.tchorzyksen.my.web.service;
+
+import org.springframework.boot.test.context.TestConfiguration;
+
+@TestConfiguration
+public class FunctionalTestConfiguration {
+}
