--- src/main/java/pl/tchorzyksen/SpringApplicationContext.java
@@ -1,19 +1,0 @@
-package pl.tchorzyksen;
-
-import org.springframework.beans.BeansException;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationContextAware;
-
-public class SpringApplicationContext implements ApplicationContextAware {
-
-  private static ApplicationContext CONTEXT;
-
-  @Override
-  public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
-    CONTEXT = applicationContext;
-  }
-
-  public static Object getBean(String beanName) {
-    return CONTEXT.getBean(beanName);
-  }
-}
