--- src/main/java/pl/tchorzyksen/entity/PersonEntity.java
@@ -1,20 +1,0 @@
-package pl.tchorzyksen.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import lombok.Data;
-
-@Data
-@Entity(name = "person")
-public class PersonEntity {
-
-  @Id @GeneratedValue private long id;
-
-  @Column(name = "first_name", nullable = false, length = 50)
-  private String firstName;
-
-  @Column(name = "last_name", nullable = false, length = 50)
-  private String lastName;
-}
