+++ src/main/java/pl/tchorzyksen/my/web/service/SpringApplicationContext.java
@@ -1,0 +1,19 @@
+package pl.tchorzyksen.my.web.service;
+
+import org.springframework.beans.BeansException;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.ApplicationContextAware;
+
+public class SpringApplicationContext implements ApplicationContextAware {
+
+  private static ApplicationContext CONTEXT;
+
+  @Override
+  public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
+    CONTEXT = applicationContext;
+  }
+
+  public static Object getBean(String beanName) {
+    return CONTEXT.getBean(beanName);
+  }
+}
