+++ src/main/java/pl/tchorzyksen/my/web/service/entities/GreetingEntity.java
@@ -1,0 +1,5 @@
+package pl.tchorzyksen.my.web.service.entities;
+
+public record GreetingEntity(long id, String content) {
+
+}
