+++ src/main/java/pl/tchorzyksen/my/web/service/entities/PersonEntity.java
@@ -1,0 +1,22 @@
+package pl.tchorzyksen.my.web.service.entities;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import lombok.Data;
+
+@Data
+@Entity(name = "person")
+public class PersonEntity {
+
+  @Id
+  @GeneratedValue
+  private long id;
+
+  @Column(name = "first_name", nullable = false, length = 50)
+  private String firstName;
+
+  @Column(name = "last_name", nullable = false, length = 50)
+  private String lastName;
+}
