+++ src/main/java/pl/tchorzyksen/my/web/service/service/UserService.java
@@ -1,0 +1,11 @@
+package pl.tchorzyksen.my.web.service.service;
+
+import org.springframework.security.core.userdetails.UserDetailsService;
+import pl.tchorzyksen.my.web.service.model.dto.UserDto;
+
+public interface UserService extends UserDetailsService {
+
+  UserDto createUser(UserDto userDto);
+
+  UserDto getUser(String email);
+}
