--- src/main/java/pl/tchorzyksen/security/AppProperties.java
@@ -1,16 +1,0 @@
-package pl.tchorzyksen.security;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.env.Environment;
-import org.springframework.stereotype.Component;
-
-@Component
-public class AppProperties {
-
-    @Autowired
-    private Environment environment;
-
-    public String getTokeSecret() {
-    return environment.getProperty("tokenSecret");
-  }
-}
