--- src/main/java/pl/tchorzyksen/service/UserService.java
@@ -1,11 +1,0 @@
-package pl.tchorzyksen.service;
-
-import org.springframework.security.core.userdetails.UserDetailsService;
-import pl.tchorzyksen.shared.dto.UserDto;
-
-public interface UserService extends UserDetailsService {
-
-  UserDto createUser(UserDto userDto);
-
-  UserDto getUser(String email);
-}
