--- src/main/java/pl/tchorzyksen/shared/dto/PersonDto.java
@@ -1,13 +1,0 @@
-package pl.tchorzyksen.shared.dto;
-
-import lombok.Data;
-
-@Data
-public class PersonDto {
-
-  private long id;
-
-  private String firstName;
-
-  private String lastName;
-}
