--- src/main/java/pl/tchorzyksen/ui/model/request/PersonRequestModel.java
@@ -1,11 +1,0 @@
-package pl.tchorzyksen.ui.model.request;
-
-import lombok.Data;
-
-@Data
-public class PersonRequestModel {
-
-  private String firstName;
-
-  private String lastName;
-}
