--- src/main/java/pl/tchorzyksen/ui/model/request/UserLoginRequestModel.java
@@ -1,13 +1,0 @@
-package pl.tchorzyksen.ui.model.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserLoginRequestModel {
-
-  private String email;
-
-  private String password;
-}
