--- src/main/java/pl/tchorzyksen/ui/model/request/UserRequestModel.java
@@ -1,15 +1,0 @@
-package pl.tchorzyksen.ui.model.request;
-
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
-@Data
-public class UserRequestModel {
-
-  private PersonRequestModel person;
-
-  private String email;
-
-  private String password;
-}
