--- src/main/java/pl/tchorzyksen/ui/model/response/PersonResponse.java
@@ -1,13 +1,0 @@
-package pl.tchorzyksen.ui.model.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class PersonResponse {
-
-  private String firstName;
-
-  private String lastName;
-}
